reorganizing file refactoring the windows add files into split/add modes
