+++ CMakeLists.txt
@@ -9,9 +9,9 @@ add_library(libeasylogs STATIC IMPORTED)
 set_property(TARGET libeasylogs PROPERTY IMPORTED_LOCATION ${CMAKE_SOURCE_DIR}/Libraries/libeasylogs.a)
 include_directories(${CMAKE_SOURCE_DIR}/Headers)
 
-set(SOURCES Sources/main.c Sources/Init/create_entities.c Headers/create_entities.h Sources/Init/init_mask.c Sources/Instances/push_back.c Sources/Instances/delete.c Sources/Instances/window_manage.c Sources/Properties/states.c Headers/properties.h)
+set(SOURCES Sources/main.c Sources/Init/create_entities.c Headers/create_entities.h Sources/Init/init_mask.c Sources/Instances/push_back.c Sources/Instances/delete.c Sources/Instances/window_manage.c Sources/Properties/states.c Headers/properties.h Sources/Windows/Dropdown/dropdown_manage.c Sources/Instances/window_split.c Sources/Instances/windows_add.c)
 
-set(C_FILES ${SOURCES})
+set(C_FILES ${SOURCES} Headers/window_manage.h)
 message(STATUS "Creating sources from: ${C_FILES}")
 add_executable(SimpleGui ${C_FILES})
